+++ src/core/CFUtils.gd
@@ -210,3 +210,14 @@ static func compare_strings(s1: String, s2: String, comparison_type: String) -> 
 			if s1 != s2:
 				comp_result = true
 	return(comp_result)
+
+static func get_unique_values(property) -> Array:
+	var unique_property_values := []
+	if property in CardConfig.PROPERTIES_STRINGS:
+		for card_def in cfc.card_definitions:
+
+			if not cfc.card_definitions[card_def][property]\
+					in unique_property_values:
+				unique_property_values.append(
+						cfc.card_definitions[card_def][property])
+	return(unique_property_values)