+++ src/core/DeckBuilder/DeckNameRandomizer.gd
@@ -1,0 +1,59 @@
+class_name DeckNameRandomizer
+extends Reference
+
+
+var adjectives = [
+	"Awesome", "Brilliant", "New", "Newer", "Newest", "Stable",
+	"Solid", "Jank", "Jankiest", "Questionable", "Curious", "Weird",
+	"Broken", "Overpowered", "Fun", "Fast", "Ephemeral", "Blitz",
+	"Cautious", "Control", "Scientific", "Explosive", "Poisonous", "Humorous",
+	"Green", "Crimson", "Yellow", "Black", "White", "Terracota", "Blue", "Grey",
+	"Holy", "Unholy", "Mighty", "Weak", "Boring", "Swimming", "Drowning",
+	"Ridiculous", "Embarasing", "Discombobulating", "Routing", "Boring", 
+	"Working", "Failing", "Ugly", "Bad", "Good", "Horrible", "Terrible", 
+	"Catalytic", "Mind-blowing", "Cute", "Slow", "Elaborate", "Cracked", 
+	"Cheesy", "Rich", "Poor", "Alchemical", "Squishy", "Lagging", "Stumbling",
+	"Horny", "Wooden", "Polished", "Sharp", "Asynchronous", "Phasing",
+	"Beholding", "Reasonable", "Absurd", "Irrational", "Logical", "Spineless",
+	"Fragmented", "Sorry", "Freak", "Geeky", "Nuclear", "Radioactive",
+	"Breaking", "Magic", "Quivering", "Fungal", "Rotting", "Phyrric",
+	"Ethereal", "Astral", "Cold", "Hot", "Zodiac", "Proletarian",
+	"Conjuring", "Rapid", "Rotating", "Searing", "Evil", "Thieving",
+	"Fractal", "Grimdark", "Dank", "Desperate", "Aerodynamic", "Neutral",
+	"Heretic", "Rugose", 
+]
+var nouns = [
+	"Deck", "Deck", "Deck", "Bible", "Weapon", "Jank", "Pattern",
+	"Pile", "Array", "Garbage", "Book", "Disaster", "Destruction", "Victory",
+	"Loss", "Shuffle", "Fatality", "Discovery", "Exploration", "God", 
+	"Monster", "Potato", "Zuccini", "Broccoli", "Demon", "Angel", "Beast",
+	"Salad", "Knife", "Spork", "Soup", "Gazebo", "Pea", "Penis", "Rose",
+	"Match", "Flower", "Bullet", "Pepper", "Clown", "Clone", "Joke", "Duck",
+	"Foot", "Palm", "Cheese", "Monkey", "Gorilla", "Donkey", "Ninja", "Samurai",
+	"Bank", "Bull", "Corner", "Dollar", "Credit", "Pogo", "Music", "Melody",
+	"Souffle", "Meat", "Blitz", "Chemistry", "Pling", "Ping", "Link", "Catastrophy",
+	"Butt", "Tooth", "Eyeball", "Brain", "Erection", "Stairway", "Song", "Whale",
+	"Doplhin", "Shark", "Octopus", "Amoeba", "Proton", "Zealot", "Neutron", "Atom",
+	"Pie", "Pi", "Radius", "Circle", "Cake", "Dam", "Wave", "Sailor", "Spell",
+	"Source", "Mushroom", "Pulse", "Pus", "Andromeda", "Galaxy", "Doctor",
+	"Star", "Nebula", "Black Hole", "Sun", "Rock", "Meteor", "Zenith", "leader",
+	"Grasshopper", "Cockroach", "Mouse", "Bug", "Wasp", "Laser", "Beam", "Crystal",
+	"Sword", "Bow", "Hammer", "Axe", "Spear", "Shield", "Overlord", "Warlord",
+	"King", "Champion", "Hero", "Peasant", "Scum", "Bucket", "Descent", "Cult"
+	
+]
+var appends = [
+	"of Death", "of Madness", "of Steel", "of Ichor", "of Cards",
+	"of Silence", "of Fire", "of Water", "of Air", "of Earth", "of Fruit",
+	"of War", "of Peace", "of Sloth", "of Rage", "of Pride", "of Lust",
+	"of Salt", "of Wat", "of Worth", "of Horror", "of Terror", "of Jokes",
+	"of Marble", "of Wat", "of Cheese", "of Doom", "of Desire", "of Passion",
+	"of the Cosmos", "of the Nether", "from the Wormhole", "out of Time",
+	"of Heaven", "of Hell", "of Limbo", "from Tartarus", "of Charm", "of Lies",
+	"of Annoyance", "of Sorcery", "of Flesh", "of Clarity", "of Ice", "of Lightning",
+	"of the Serpent", "of Nothing", "of the People", "of the Master", "of the Noob",
+	"of the Elements", "of Malice", "of the Underworld", "of Wonders", "of Sanity",
+	"of the Void", "of Chaos", "of Order", "of the Elder Gods", "of Blasphemy"
+	
+	
+]